final_package/requirements.txt
pandas
numpy
openpyxl
requests
msal
azure-functions
